+++ src/encoded/types/user_content.py
@@ -159,7 +159,7 @@ class StaticSection(UserContent):
         elif file_type == 'md':
             # remove new line character
             output = convert_markdown_to_html(content)
-            output = output.replace('\n', '')
+            # output = output.replace('\n', '')
             if output and convert_ext_links:
                 return convert_external_links(output, request.domain)
             return output
@@ -350,7 +350,7 @@ def get_remote_file_contents(uri):
 
 def convert_markdown_to_html(markdown_text, custom_wrapper = 'div'):
     # convert markdown to html including tables
-    html_output = markdown.markdown(markdown_text, extensions=['tables'])
+    html_output = markdown.markdown(markdown_text, extensions=['tables', 'fenced_code'])
 
     # check content has any header, if yes wrap it with custom tag
     header_pattern = re.compile(r'<h[1-6]>.*?<\/h[1-6]>', re.IGNORECASE)